+++ sources/dicom/large_image_source_dicom/__init__.py
@@ -1,3 +1,4 @@
+import functools
 import math
 import os
 import re
@@ -93,9 +94,30 @@ def dicom_to_dict(ds, base=None):
     return info
 
 
+@functools.lru_cache(maxsize=10000)
+def _getSeriesUIDForPath(path):
+    """
+    Check if the current path can be opened via dicom and returns a
+    SeriesInstanceUID.
+
+    :param path: a path to a potential dicom.
+    :returns: the SeriesInstanceUID if present; None if not or not a DICOM.
+    """
+    _lazyImportPydicom()
+
+    try:
+        series_uid = pydicom.filereader.dcmread(
+            path, stop_before_pixels=True, specific_tags=['SeriesInstanceUID'],
+        )[pydicom.tag.Tag('SeriesInstanceUID')].value
+        return series_uid
+    except Exception:
+        return None
+
+
 class DICOMFileTileSource(FileTileSource, metaclass=LruCacheMetaclass):
     """
-    Provides tile access to dicom files the dicom or dicomreader library can read.
+    Provides tile access to dicom files the dicom or dicomreader library can
+    read.
     """
 
     cacheName = 'tilesource'
@@ -246,24 +268,10 @@ class DICOMFileTileSource(FileTileSource, metaclass=LruCacheMetaclass):
         if not mightbe and re.match(r'^DCM_\d+$', path):
             mightbe = True
         if mightbe and basepath:
-            try:
-                _lazyImportPydicom()
-                base = pydicom.filereader.dcmread(basepath, stop_before_pixels=True)
-            except Exception as exc:
-                msg = f'File cannot be opened via dicom tile source ({exc}).'
-                raise TileSourceError(msg)
-            try:
-                base_series_uid = base[pydicom.tag.Tag('SeriesInstanceUID')].value
-            except Exception:
-                base_series_uid = None
+            base_series_uid = _getSeriesUIDForPath(basepath)
             if base_series_uid:
-                try:
-                    series_uid = pydicom.filereader.dcmread(
-                        origpath, stop_before_pixels=True,
-                    )[pydicom.tag.Tag('SeriesInstanceUID')].value
-                    mightbe = base_series_uid == series_uid
-                except Exception:
-                    mightbe = False
+                series_uid = _getSeriesUIDForPath(origpath)
+                mightbe = series_uid is not None and base_series_uid == series_uid
         return mightbe
 
     def getNativeMagnification(self):